Remove all deprecated block data values

+++ src/main/java/me/eccentric_nz/TARDIS/control/TARDISRandomButton.java
@@ -94,7 +94,7 @@ public class TARDISRandomButton {
                 String environment = "THIS";
                 int nether_min = plugin.getArtronConfig().getInt("nether_min");
                 int the_end_min = plugin.getArtronConfig().getInt("the_end_min");
-                byte[] repeaters = rsr.getRepeaters();
+                int[] repeaters = rsr.getRepeaters();
                 if (repeaters[0] == -1) {
                     TARDISMessage.send(player, "FLIGHT_BAD");
                     return;
